+++ UIKitSwagger/View/ViewCorners.swift
@@ -6,7 +6,7 @@
 //  Copyright © 2016 Swagger Soft. All rights reserved.
 //
 
-import Foundation
+import UIKit
 
 public struct ViewCorners: OptionSetType {