+++ UIKitSwagger/build-docs.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+PROJECT_DIR = ../
+OUTPUT_DIR = $PROJECT_DIR/Docs/api