+++ UIKitSwaggerTests/Tools/Common/MethodSwizzler.m
@@ -1,0 +1,13 @@
+//
+//  MethodSwizzler.m
+//  UIKitSwagger
+//
+//  Created by Sam Odom on 4/29/16.
+//  Copyright © 2016 Swagger Soft. All rights reserved.
+//
+
+#import "MethodSwizzler.h"
+
+@implementation MethodSwizzler
+
+@end